pyproject.toml
[tool.poetry]
name = "count-web-content"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "count_web_content" }]


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
